Fix issue with `undefined` in encrypted text view
Issue
=====
An issue with view happened when user tried to encrypt characters which
do not belong to defined charlist. If there were two or more such
characters in a row, the 'undefined' word appeared in the view field.

To reproduce the issue you can try to encrypt some cyrillic text. For
example you can try this: "Привіт, світе!" (which means "Hello world!")

Problem
-------
The problem is hidden here: https://github.com/mozilla/codemoji/blob/master/assets/js/ui.js#L151

When you try to display a text with unknown characters like "Привіт,
світе!" it will be transformed with twemoji function into something like
this: "Привіт<img ...> світе<img ...>". After regexp_replace and split
occurs that is transformed into the list of two elements:

.. code::
    ["Привіт<img ...>", " світе<img ...>"]

And when you try to access 3,4,5, and so forth element on line 155:
https://github.com/mozilla/codemoji/blob/master/assets/js/ui.js#L155 it
returnes undefined which is displayed later.

P.S. I may make some mistakes about handling whitespaces, for example,
but the problem in general is the same as I described.

Solution
--------
The solution is to find a better way to split that twemojified string
into elements.

Commit
======
This commit provides a solution to the described issue. As there is
jQuery loaded into that module, I use it to split twemojified string
into the list of DOM Elements. If the element is an Element node, I use
its string representation to build the output. Otherwise I assume that
it is a Text node. I split it character by character and add each
character as separate character to splitted elements list.

Than I use that list to build final result